+++ JudoPayDotNetTests/Clients/PaymentTests.cs
@@ -55,7 +55,6 @@ namespace JudoPayDotNetTests.Clients
                             currency : 'GBP',
                             consumer : 
                                 {
-                                    consumerToken : 'B245SEB',
                                     yourConsumerReference : 'Consumer1'
                                 }
                             }",
@@ -93,7 +92,6 @@ namespace JudoPayDotNetTests.Clients
                             currency : 'GBP',
                             consumer :
                                 {
-                                    consumerToken : 'B245SEB',
                                     yourConsumerReference : 'Consumer1'
                                 }
                             }",
@@ -109,8 +107,7 @@ namespace JudoPayDotNetTests.Clients
                                                  JudoId = "100200300",
                                                  MobileNumber = "07999999999",
                                                  PhoneCountryCode = "44",
-                                                 YourConsumerReference = "User10",
-                                                 ConsumerToken = "ABAS"
+                                                 YourConsumerReference = "User10"
                                              },
                             @"{
                             receiptId : '134567',
@@ -129,7 +126,6 @@ namespace JudoPayDotNetTests.Clients
                             currency : 'GBP',
                             consumer : 
                                 {
-                                    consumerToken : 'B245SEB',
                                     yourConsumerReference : 'Consumer1'
                                 }
                             }",
@@ -180,8 +176,7 @@ namespace JudoPayDotNetTests.Clients
                                                  JudoId = "100200300",
                                                  MobileNumber = "07999999999",
                                                  PhoneCountryCode = "44",
-                                                 YourConsumerReference = "User10",
-                                                 ConsumerToken = "ABAS"
+                                                 YourConsumerReference = "User10"
                                              },
                             @"    
                         {