+++ JudoPayDotNetTests/Clients/PreAuthTests.cs
@@ -62,7 +62,6 @@ namespace JudoPayDotNetTests.Clients
                             currency : 'GBP',
                             consumer : 
                                 {
-                                    consumerToken : 'B245SEB',
                                     yourConsumerReference : 'Consumer1'
                                 }
                             }",
@@ -81,9 +80,7 @@ namespace JudoPayDotNetTests.Clients
                         JudoId = "100200300",
                         MobileNumber = "07999999999",
                         PhoneCountryCode = "44",
-                        YourConsumerReference = "User10",
-                       
-                        ConsumerToken = "ABAS"
+                        YourConsumerReference = "User10"
                     },
                         @"{
                             receiptId : '134567',
@@ -165,8 +162,7 @@ namespace JudoPayDotNetTests.Clients
                         JudoId = "100200300",
                         MobileNumber = "07999999999",
                         PhoneCountryCode = "44",
-                        YourConsumerReference = "User10",
-                        ConsumerToken = "ABAS"
+                        YourConsumerReference = "User10"
                     },
                         @"    
                         {